+++ bin/server.js
@@ -37,7 +37,6 @@ app.use(Logger.requests(Log, Config.get('log:level')));
 
 // Add middleware for paring JSON requests
 app.use(BodyParser.json());
-app.use(BodyParser.urlencoded({extended: false}));
 
 app.use(CookieParser());
