+++ lib/control/v1/index.js
@@ -1,13 +1,13 @@
 'use strict';
 
 const Err = require('./error');
+const upload = require('multer')();
 
 module.exports = function Index(app) {
   app.get('/', function(req, res, next) {
     res.render('index', {title: 'ACS'});
   });
 
-  app.post('/encrypt/get_ct_json', require('./encrypt')());
-
+  app.post('/v1/vault', upload.none(), require('./vault')());
   app.use(Err);
 };
