+++ lib/control/v1/vault.js
