+++ package.json
@@ -23,6 +23,7 @@
     "ejs": "^2.5.5",
     "express": "^4.14.1",
     "express-winston": "^2.0.0",
+    "multer": "^1.3.0",
     "nconf": "^0.8.4",
     "request": "^2.75.0",
     "request-promise": "^4.1.1",
